Made UI changes and bug fixing
