+++ api/app/db/seeds.py
@@ -59,7 +59,7 @@ def seed_roles():
 def seed_admin_user():
     """Create the default admin user"""
     admin_email = "admin@jothis.com"
-    admin_password = "J@Th&s!@178’/"
+    admin_password = "J@Th&s!@178/"
     
     # Check if admin already exists
     existing_user = users.find_one({"email": admin_email})
